+++ R/get_eigenvalues.R
@@ -18,3 +18,20 @@ setMethod(
     z
   }
 )
+
+#' @export
+#' @rdname get_eigenvalues
+#' @aliases get_eigenvalues,PCOA-method
+setMethod(
+  f = "get_eigenvalues",
+  signature = c(x = "PCOA"),
+  definition = function(x) {
+    eig <- x@eigenvalues# Eigenvalues
+    pvar <- eig / sum(eig) * 100 # Percentage
+    cvar <- cumsum(pvar) # Cumulative percentage
+
+    z <- data.frame(eig, pvar, cvar)
+    colnames(z) <- c("eigenvalues", "variance", "cumulative")
+    z
+  }
+)