+++ R/viz_coordinates.R
@@ -10,17 +10,18 @@ setMethod(
   f = "viz_rows",
   signature = c(x = "MultivariateAnalysis"),
   definition = function(x, ..., axes = c(1, 2), active = TRUE, sup = TRUE,
-                        labels = FALSE, highlight = NULL,
-                        color = NULL, symbol = FALSE, size = c(1, 6),
+                        labels = FALSE, extra_quali = NULL, extra_quanti = NULL,
+                        color = NULL, fill = FALSE, symbol = FALSE, size = c(1, 6),
                         xlim = NULL, ylim = NULL, main = NULL, sub = NULL,
                         panel.first = NULL, panel.last = NULL,
                         legend = list(x = "topleft")) {
-    viz_points(x, margin = 1, axes = axes, active = active, sup = sup,
-               labels = labels, highlight = highlight,
-               color = color, symbol = symbol, size = size,
+    viz_points(x, margin = 1, axes = axes, ...,
+               active = active, sup = sup, labels = labels,
+               extra_quali = extra_quali, extra_quanti = extra_quanti,
+               color = color, fill = fill, symbol = symbol, size = size,
                xlim = xlim, ylim = ylim, main = main, sub = sub,
                panel.first = panel.first, panel.last = panel.last,
-               legend = legend, ...)
+               legend = legend)
     invisible(x)
   }
 )
@@ -31,10 +32,11 @@ setMethod(
 setMethod(
   f = "viz_rows",
   signature = c(x = "BootstrapCA"),
-  definition = function(x, ..., axes = c(1, 2), color = FALSE, symbol = FALSE) {
-    group <- get_groups(x, margin = 1)
-    viz_points(x, ..., margin = 1, axes = axes, active = TRUE, sup = TRUE,
-               labels = FALSE, highlight = group, color = color, symbol = symbol)
+  definition = function(x, ..., axes = c(1, 2), color = FALSE, fill = FALSE,
+                        symbol = FALSE) {
+    viz_points(x, margin = 1, axes = axes, ..., active = TRUE, sup = TRUE,
+               labels = FALSE, extra_quali = NULL,
+               color = color, fill = fill, symbol = symbol)
     invisible(x)
   }
 )
@@ -47,17 +49,18 @@ setMethod(
   f = "viz_individuals",
   signature = c(x = "PCA"),
   definition = function(x, ..., axes = c(1, 2), active = TRUE, sup = TRUE,
-                        labels = FALSE, highlight = NULL,
-                        color = NULL, symbol = FALSE, size = c(1, 6),
+                        labels = FALSE, extra_quali = NULL, extra_quanti = NULL,
+                        color = NULL, fill = FALSE, symbol = FALSE, size = c(1, 6),
                         xlim = NULL, ylim = NULL, main = NULL, sub = NULL,
                         panel.first = NULL, panel.last = NULL,
                         legend = list(x = "topleft")) {
-    viz_points(x, margin = 1, axes = axes, active = active, sup = sup,
-               labels = labels, highlight = highlight,
-               color = color, symbol = symbol, size = size,
+    viz_points(x, margin = 1, axes = axes, ...,
+               active = active, sup = sup, labels = labels,
+               extra_quali = extra_quali, extra_quanti = extra_quanti,
+               color = color, fill = fill, symbol = symbol, size = size,
                xlim = xlim, ylim = ylim, main = main, sub = sub,
                panel.first = panel.first, panel.last = panel.last,
-               legend = legend, ...)
+               legend = legend)
     invisible(x)
   }
 )
@@ -70,17 +73,18 @@ setMethod(
   f = "viz_columns",
   signature = c(x = "MultivariateAnalysis"),
   definition = function(x, ..., axes = c(1, 2), active = TRUE, sup = TRUE,
-                        labels = FALSE, highlight = NULL,
-                        color = NULL, symbol = FALSE, size = c(1, 6),
+                        labels = FALSE, extra_quali = NULL, extra_quanti = NULL,
+                        color = NULL, fill = FALSE, symbol = FALSE, size = c(1, 6),
                         xlim = NULL, ylim = NULL, main = NULL, sub = NULL,
                         panel.first = NULL, panel.last = NULL,
                         legend = list(x = "topleft")) {
-    viz_points(x, margin = 2, axes = axes, active = active, sup = sup,
-               labels = labels, highlight = highlight,
-               color = color, symbol = symbol, size = size,
+    viz_points(x, margin = 2, axes = axes, ...,
+               active = active, sup = sup, labels = labels,
+               extra_quali = extra_quali, extra_quanti = extra_quanti,
+               color = color, fill = fill, symbol = symbol, size = size,
                xlim = xlim, ylim = ylim, main = main, sub = sub,
                panel.first = panel.first, panel.last = panel.last,
-               legend = legend, ...)
+               legend = legend)
     invisible(x)
   }
 )
@@ -91,10 +95,11 @@ setMethod(
 setMethod(
   f = "viz_columns",
   signature = c(x = "MultivariateBootstrap"),
-  definition = function(x, ..., axes = c(1, 2), color = FALSE, symbol = FALSE) {
-    group <- get_groups(x, margin = 2)
+  definition = function(x, ..., axes = c(1, 2), color = FALSE, fill = FALSE,
+                        symbol = FALSE) {
     viz_points(x, ..., margin = 2, axes = axes, active = TRUE, sup = TRUE,
-               labels = FALSE, highlight = group, color = color, symbol = symbol)
+               labels = FALSE, extra_quali = NULL,
+               color = color, fill = fill, symbol = symbol)
     invisible(x)
   }
 )
@@ -107,17 +112,17 @@ setMethod(
   f = "viz_variables",
   signature = c(x = "PCA"),
   definition = function(x, ..., axes = c(1, 2), active = TRUE, sup = TRUE,
-                        labels = list(how = "contribution", n = 10),
-                        highlight = NULL,
+                        labels = list(filter = "contribution", n = 10),
+                        extra_quali = NULL, extra_quanti = NULL,
                         color = NULL, symbol = NULL, size = 1,
                         xlim = NULL, ylim = NULL, main = NULL, sub = NULL,
                         panel.first = NULL, panel.last = NULL,
                         legend = list(x = "topleft")) {
     ## Prepare data
-    coord <- prepare(x, margin = 2, axes = axes, active = active,
-                     sup = sup, highlight = highlight,
-                     color = color, line_type = symbol,
-                     line_width = size, ...)
+    coord <- prepare_plot(x, margin = 2, axes = axes, ...,
+                          active = active, sup = sup,
+                          extra_quali = extra_quali, extra_quanti = extra_quanti,
+                          color = color, line_type = symbol, line_width = size)
 
     ## Save and restore graphical parameters
     ## pty: square plotting region, independent of device size
@@ -159,7 +164,7 @@ setMethod(
     ## Labels
     if (isTRUE(labels)) labels <- list()
     if (is.list(labels)) {
-      viz_labels(coord, select = labels$how, n = labels$n)
+      viz_labels(coord, filter = labels$filter, n = labels$n)
     }
 
     ## Evaluate post-plot and pre-axis expressions
@@ -199,17 +204,18 @@ setMethod(
   f = "viz_variables",
   signature = c(x = "CA"),
   definition = function(x, ..., axes = c(1, 2), active = TRUE, sup = TRUE,
-                        labels = FALSE, highlight = NULL,
-                        color = NULL, symbol = FALSE, size = c(1, 6),
+                        labels = FALSE, extra_quali = NULL, extra_quanti = NULL,
+                        color = NULL, fill = FALSE, symbol = FALSE, size = c(1, 6),
                         xlim = NULL, ylim = NULL, main = NULL, sub = NULL,
                         panel.first = NULL, panel.last = NULL,
                         legend = list(x = "topleft")) {
-    viz_points(x, margin = 2, axes = axes, active = active, sup = sup,
-               labels = labels, highlight = highlight,
-               color = color, symbol = symbol, size = size,
+    viz_points(x, margin = 2, axes = axes, ...,
+               active = active, sup = sup, labels = labels,
+               extra_quali = extra_quali, extra_quanti = extra_quanti,
+               color = color, fill = fill, symbol = symbol, size = size,
                xlim = xlim, ylim = ylim, main = main, sub = sub,
                panel.first = panel.first, panel.last = panel.last,
-               legend = legend, ...)
+               legend = legend)
   }
 )
 
@@ -219,10 +225,11 @@ setMethod(
 setMethod(
   f = "viz_variables",
   signature = c(x = "BootstrapPCA"),
-  definition = function(x, ..., axes = c(1, 2), color = FALSE, symbol = FALSE) {
-    group <- get_groups(x, margin = 2)
+  definition = function(x, ..., axes = c(1, 2), color = FALSE, fill = FALSE,
+                        symbol = FALSE) {
     viz_points(x, ..., margin = 2, axes = axes, active = TRUE, sup = TRUE,
-               labels = FALSE, highlight = group, color = color, symbol = symbol)
+               labels = FALSE, extra_quali = NULL,
+               color = color, fill = fill, symbol = symbol)
     invisible(x)
   }
 )
@@ -234,9 +241,9 @@ setMethod(
 #' @param labels A [`logical`] scalar: should labels be drawn? Labeling a large
 #'  number of points can be computationally expensive and make the graph
 #'  difficult to read. A selection of points to label can be provided using a
-#'  `list` of two named elements, `how` (a string specifying how to select the
-#'  labels to be drawn) and `n` (an integer specifying the number of labels to
-#'  be drawx). See examples below.
+#'  `list` of two named elements, `filter` (a string specifying how to filter
+#'  the labels to be drawn) and `n` (an integer specifying the number of labels
+#'  to be drawn). See examples below.
 #' @param xlim A length-two [`numeric`] vector giving the x limits of the plot.
 #'  The default value, `NULL`, indicates that the range of the
 #'  [finite][is.finite()] values to be plotted should be used.
@@ -259,14 +266,14 @@ setMethod(
 #'  [graphics::legend()]; names of the list are used as argument names.
 #'  If `NULL`, no legend is displayed.
 #' @param ... Currently not used.
-#' @param symbol A vector of plotting characters or symbols.
-#' @inheritParams prepare
+#' @inheritParams prepare_plot
 #' @author N. Frerebeau
 #' @keywords internal
 viz_points <- function(x, margin, axes, ...,
                        active = TRUE, sup = TRUE,
-                       labels = list(how = "contribution", n = 10),
-                       highlight = NULL, color = NULL,
+                       labels = list(filter = "contribution", n = 10),
+                       extra_quali = NULL, extra_quanti = NULL,
+                       color = NULL, fill = FALSE,
                        symbol = NULL, size = c(1, 6),
                        xlim = NULL, ylim = NULL,
                        main = NULL, sub = NULL, xlab = NULL, ylab = NULL,
@@ -274,9 +281,11 @@ viz_points <- function(x, margin, axes, ...,
                        panel.first = NULL, panel.last = NULL,
                        legend = list(x = "topleft")) {
   ## Prepare data
-  coord <- prepare(x, margin = margin, axes = axes, active = active,
-                   sup = sup, highlight = highlight,
-                   color = color, shape = symbol, size = size, ...)
+  coord <- prepare_plot(x, margin = margin, axes = axes,
+                        active = active, sup = sup,
+                        extra_quali = extra_quali, extra_quanti = extra_quanti,
+                        color = color, fill = fill,
+                        symbol = symbol, size = size, ...)
 
   ## Save and restore graphical parameters
   ## pty: square plotting region, independent of device size
@@ -311,7 +320,7 @@ viz_points <- function(x, margin, axes, ...,
   ## Labels
   if (isTRUE(labels)) labels <- list()
   if (is.list(labels)) {
-    viz_labels(coord, select = labels$how, n = labels$n)
+    viz_labels(coord, filter = labels$filter, n = labels$n)
   }
 
   ## Evaluate post-plot and pre-axis expressions
@@ -343,18 +352,13 @@ viz_points <- function(x, margin, axes, ...,
   invisible(coord)
 }
 
-print_variance <- function(object, axis) {
-  v <- get_variance(object, digits = 1) # Get percentage of variance
-  sprintf("%s (%g%%)", names(v)[[axis]], v[[axis]])
-}
-
 #' Non-Overlapping Text Labels
 #'
-#' @param x A [`data.frame`] (typically returned by [prepare()]).
-#' @param select A [`character`] string specifying the variable to select
-#'  labels. If `NULL`, all labels are drawn.
+#' @param x A [`data.frame`] (typically returned by [prepare_plot()]).
+#' @param filter A [`character`] string specifying the variable used to filter
+#'  observations. If `NULL`, all labels are drawn.
 #' @param n An [`integer`] specifying the number of labels to draw.
-#'  Only the labels of the top \eqn{n} observations according to `select` will
+#'  Only the labels of the top \eqn{n} observations according to `filter` will
 #'  be drawn. If `NULL`, all labels are drawn.
 #' @param type A [`character`] string specifying the shape of the field.
 #'  It must be one of "`text`", "`shadow`" or "`box`". Any unambiguous substring
@@ -362,12 +366,12 @@ print_variance <- function(object, axis) {
 #' @param ... Currently not used.
 #' @author N. Frerebeau
 #' @keywords internal
-viz_labels <- function(x, select = "contribution", n = 10,
+viz_labels <- function(x, filter = "contribution", n = 10,
                        type = "shadow", ...) {
   ## Select
-  if (!is.null(select) && !is.null(n) && n > 0) {
+  if (!is.null(filter) && !is.null(n) && n > 0) {
     top <- min(nrow(x), n)
-    how <- x[[select]]
+    how <- x[[filter]]
     k <- order(how, decreasing = TRUE)[seq_len(top)] # Get order
     x <- x[k, , drop = FALSE] # Subset
   }
@@ -381,4 +385,210 @@ viz_labels <- function(x, select = "contribution", n = 10,
     # cex = x$cex,
     xpd = TRUE
   )
+}
+
+#' Prepare Data for Plotting
+#'
+#' @param x A [`MultivariateAnalysis-class`] object.
+#' @param margin A length-one [`numeric`] vector giving the subscript
+#'  which the data will be returned: `1` indicates individuals/rows (the
+#'  default), `2` indicates variables/columns.
+#' @param axes A length-two [`numeric`] vector giving the dimensions to be
+#'  plotted.
+#' @param active A [`logical`] scalar: should the active observations be
+#'  plotted?
+#' @param sup A [`logical`] scalar: should the supplementary observations be
+#'  plotted?
+#' @param principal A [`logical`] scalar: should principal coordinates be
+#'  returned? If `FALSE`, standard coordinates are returned.
+#' @param extra_quali An optional vector of qualitative data for aesthetics
+#'  mapping.
+#' @param extra_quanti An optional vector of quantitative data for aesthetics
+#'  mapping. If a single [`character`] string is passed, it must be one of
+#'  "`observation`", "`mass`", "`sum`", "`contribution`" or "`cos2`"
+#'  (see [`augment()`]).
+#' @param color The colors for lines and points (will be mapped to
+#'  `extra_quanti` or `extra_quali`; if both are set, the latter has priority).
+#'  Ignored if set to `FALSE`.
+#' @param fill The background colors for points (will be mapped to
+#'  `extra_quanti` or `extra_quali`; if both are set, the latter has priority).
+#'  Ignored if set to `FALSE`.
+#' @param symbol A vector of plotting characters or symbols (will be mapped to
+#'  `extra_quali`). This can either be a single character or an integer code for
+#'  one of a set of graphics symbols. If `symbol` is a named a named vector,
+#'  then the symbols will be associated with their name within `extra_quali`.
+#'  Ignored if set to `FALSE`.
+#' @param size A length-two [`numeric`] vector giving range of possible sizes
+#'  (greater than 0; will be mapped to `extra_quanti`).
+#'  Ignored if set to `FALSE`.
+#' @param line_type A specification for the line type (will be mapped to
+#'  `extra_quali`). If `line_type` is a named a named vector, then the line
+#'  types will be associated with their name within `extra_quali`.
+#'  Ignored if set to `FALSE`.
+#' @param line_width A specification for the line type and width (will
+#'  be mapped to `extra_quanti`).
+#'  Ignored if set to `FALSE`.
+#' @param ... Further [graphical parameters][graphics::par].
+#' @return
+#'  A [`data.frame`] with the following columns:
+#'    \describe{
+#'     \item{`x`}{Coordinates along x.}
+#'     \item{`y`}{Coordinates along y.}
+#'     \item{`extra_quali`}{Extra qualitative variable to be highlighted.}
+#'     \item{`extra_quanti`}{Extra quantitative variable to be highlighted.}
+#'     \item{`label`}{Label.}
+#'     \item{`sup`}{Is supplementary?}
+#'     \item{`col`}{Color for lines and symbols.}
+#'     \item{`bg`}{Background color for symbols.}
+#'     \item{`pch`}{Symbols.}
+#'     \item{`cex`}{Symbol sizes.}
+#'     \item{`lty`}{Line types.}
+#'     \item{`lwd`}{Line widths.}
+#'    }
+#' @author N. Frerebeau
+#' @keywords internal
+prepare_plot <- function(x, margin, ..., axes = c(1, 2), active = TRUE,
+                         sup = TRUE, principal = TRUE,
+                         extra_quali = NULL, extra_quanti = NULL,
+                         color = NULL, fill = FALSE,
+                         symbol = NULL, size = c(1, 6),
+                         line_type = NULL, line_width = size) {
+  ## /!\ Backward compatibility /!\
+  high <- list(...)$highlight
+  if (length(high) == 1) {
+    if (high == "observation") extra_quali <- high else extra_quanti <- high
+  }
+
+  ## Prepare data
+  data <- augment(x, margin = margin, axes = axes, principal = principal)
+  n <- nrow(data)
+
+  ## Recode
+  data$observation <- ifelse(data$supplementary, "suppl.", "active")
+
+  ## Recycle graphical parameters if of length one
+  dots <- list(...)
+  col <- recycle(dots$col %||% graphics::par("col"), n)
+  bg <- recycle(dots$bg %||% graphics::par("bg"), n)
+  pch <- recycle(dots$pch %||% 16, n)
+  cex <- recycle(dots$cex %||% graphics::par("cex"), n)
+  lty <- recycle(dots$lty %||% graphics::par("lty"), n)
+  lwd <- recycle(dots$lwd %||% graphics::par("lwd"), n)
+
+  ## Highlight quantitative information
+  if (length(extra_quanti) == 1) {
+    choices <- c("mass", "sum", "contribution", "cos2")
+    extra_quanti <- match.arg(extra_quanti, choices = choices, several.ok = FALSE)
+    extra_quanti <- data[[extra_quanti]]
+  }
+  if (length(extra_quanti) > 0) {
+    arkhe::assert_length(extra_quanti, n)
+    ## Continuous scales
+    if (!isFALSE(color)) col <- khroma::palette_color_continuous(colors = color)(extra_quanti)
+    if (!isFALSE(fill)) bg <- khroma::palette_color_continuous(colors = fill)(extra_quanti)
+    if (!isFALSE(size)) cex <- khroma::palette_size_range(range = size)(extra_quanti)
+    if (!isFALSE(line_width)) lwd <- khroma::palette_size_range(range = line_width)(extra_quanti)
+  } else {
+    extra_quanti <- rep(NA_real_, n)
+  }
+
+  ## Highlight qualitative information
+  if (is.null(extra_quali) && has_groups(x, margin = margin)) {
+    extra_quali <- get_groups(x, margin = margin)
+  }
+  if (is.character(extra_quali) && length(extra_quali) == 1) {
+    choices <- c("observation")
+    extra_quali <- match.arg(extra_quali, choices = choices, several.ok = FALSE)
+    extra_quali <- data[[extra_quali]]
+  }
+  if (!isFALSE(extra_quali) && length(extra_quali) > 0) {
+    arkhe::assert_length(extra_quali, n)
+    ## Discrete scales
+    if (!isFALSE(color)) col <- khroma::palette_color_discrete(colors = color)(extra_quali)
+    if (!isFALSE(fill)) bg <- khroma::palette_color_discrete(colors = fill)(extra_quali)
+    if (!isFALSE(symbol)) pch <- khroma::palette_shape(symbols = symbol)(extra_quali)
+    if (!isFALSE(line_type)) lty <- khroma::palette_line(types = line_type)(extra_quali)
+  } else {
+    extra_quali <- rep(NA_character_, n)
+  }
+
+  coord <- data.frame(
+    data,
+    x = data[[1L]],
+    y = data[[2L]],
+    extra_quali = extra_quali,
+    extra_quanti = extra_quanti,
+    label = data$label,
+    col = col,
+    bg = bg,
+    pch = pch,
+    cex = cex,
+    lty = lty,
+    lwd = lwd
+  )
+
+  ## Subset
+  if (active & !sup) coord <- coord[!coord$supplementary, , drop = FALSE]
+  if (!active & sup) coord <- coord[coord$supplementary, , drop = FALSE]
+
+  coord
+}
+
+#' Build a Legend
+#'
+#' @param x A [`data.frame`] returned by [prepare_plot()].
+#' @param args A [`list`] of additional arguments to be passed to
+#'  [graphics::legend()]; names of the list are used as argument names.
+#'  If `NULL` or empty, no legend is displayed.
+#' @param points A [`logical`] scalar: legend for points?
+#' @param lines A [`logical`] scalar: legend for lines?
+#' @author N. Frerebeau
+#' @keywords internal
+prepare_legend <- function(x, args, points = TRUE, lines = TRUE) {
+  quanti <- x$extra_quanti
+  quali <- x$extra_quali
+
+  if (!is.list(args) || length(args) == 0) return(NULL)
+  if (all(is.na(quanti)) && all(is.na(quali))) return(NULL)
+
+  ## Continuous scale
+  if (!all(is.na(quanti))) {
+    quanti <- quanti[!is.na(quanti)]
+    # im <- grDevices::as.raster(x$col)
+
+    pr <- pretty(quanti, n = ifelse(nrow(x) > 5, 5, nrow(x)))
+    pr <- pr[pr <= max(quanti) & pr >= min(quanti)]
+    i <- order(quanti, method = "radix")[!duplicated(quanti)]
+
+    col <- grDevices::colorRamp(x$col[i])(scale_range(pr, from = range(quanti)))
+    col <- grDevices::rgb(col, maxColorValue = 255)
+
+    leg <- list(legend = pr, col = col)
+    if (points) {
+      cex <- stats::approx(x = quanti[i], y = x$cex[i], xout = pr, ties = "ordered")$y
+      leg <- utils::modifyList(leg, list(pch = unique(x$pch), pt.cex = cex))
+    }
+    if (lines) {
+      lwd <- stats::approx(x = quanti[i], y = x$lwd[i], xout = pr, ties = "ordered")$y
+      leg <- utils::modifyList(leg, list(lty = unique(x$lty), lwd = lwd))
+    }
+  }
+  ## Discrete scale
+  if (!all(is.na(quali))) {
+    param <- stats::aggregate(
+      x[, c("col", "bg", "pch", "lty")],
+      by = list(leg = quali),
+      FUN = unique
+    )
+    leg <- list(legend = param$leg, col = param$col)
+    if (points) {
+      leg <- utils::modifyList(leg, list(pt.bg = param$bg, pch = param$pch))
+    }
+    if (lines) {
+      leg <- utils::modifyList(leg, list(lty = param$lty))
+    }
+  }
+
+  leg <- utils::modifyList(leg, args)
+  do.call(graphics::legend, args = leg)
 }